Update main.cpp
Changed shutdown command to "sudo shutdown -h now" "sudo halt" as in some cases when used with RaspBMC on Raspbian the HALT may close the RaspBMC software before Raspbmc has closed open files.  This then leaves the Pi unable to unmount the SD card as files are still open and leave it stuck and unable to power down.